[BE] Handle errors for common operations

+++ src/main/java/com/stnslv/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.stnslv.student;
 
+import com.stnslv.student.exception.BadRequestException;
+import com.stnslv.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +18,16 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
+        if (studentRepository.existsByEmail(student.getEmail())) {
+            throw new BadRequestException(String.format("Email %s is taken already", student.getEmail()));
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException("A given student is not found.");
+        }
         studentRepository.deleteById(studentId);
     }
 }
